+++ automated-tests/src/dali/utc-Dali-GestureDetector.cpp
@@ -317,7 +317,7 @@ int UtcDaliGestureDetectorDetachAllP(void)
   // Use pan gesture as GestureDetector cannot be created.
   GestureDetector detector = PanGestureDetector::New();
 
-  const unsigned int actorsToAdd = 5;
+  const size_t actorsToAdd = 5;
   std::vector<Actor> myActors;
 
   for (unsigned int i = 0; i < actorsToAdd; ++i)
@@ -332,7 +332,7 @@ int UtcDaliGestureDetectorDetachAllP(void)
   // Detach and retrieve attached actors again, the vector should be empty.
   detector.DetachAll();
 
-  DALI_TEST_EQUALS(0u, detector.GetAttachedActorCount(), TEST_LOCATION);
+  DALI_TEST_EQUALS((size_t)0u, detector.GetAttachedActorCount(), TEST_LOCATION);
   END_TEST;
 }
 
@@ -344,7 +344,7 @@ int UtcDaliGestureDetectorDetachAllN(void)
   // Use pan gesture as GestureDetector cannot be created.
   GestureDetector detector = PanGestureDetector::New();
 
-  const unsigned int actorsToAdd = 5;
+  const size_t actorsToAdd = 5;
   std::vector<Actor> myActors;
 
   for (unsigned int i = 0; i < actorsToAdd; ++i)
@@ -359,7 +359,7 @@ int UtcDaliGestureDetectorDetachAllN(void)
   // Detach and retrieve attached actors again, the vector should be empty.
   detector.DetachAll();
 
-  DALI_TEST_EQUALS(0u, detector.GetAttachedActorCount(), TEST_LOCATION);
+  DALI_TEST_EQUALS((size_t)0u, detector.GetAttachedActorCount(), TEST_LOCATION);
   DALI_TEST_EQUALS(true, gestureManager.WasCalled(TestGestureManager::UnregisterType), TEST_LOCATION);
 
   // Call DetachAll again, there should not be any exception
@@ -385,17 +385,17 @@ int UtcDaliGestureDetectorGetAttachedActors(void)
   GestureDetector detector = PanGestureDetector::New();
 
   // Initially there should not be any actors.
-  DALI_TEST_EQUALS(0u, detector.GetAttachedActorCount(), TEST_LOCATION);
+  DALI_TEST_EQUALS((size_t)0u, detector.GetAttachedActorCount(), TEST_LOCATION);
 
   // Attach one actor
   Actor actor1 = Actor::New();
   detector.Attach(actor1);
-  DALI_TEST_EQUALS(1u, detector.GetAttachedActorCount(), TEST_LOCATION);
+  DALI_TEST_EQUALS((size_t)1u, detector.GetAttachedActorCount(), TEST_LOCATION);
 
   // Attach another actor
   Actor actor2 = Actor::New();
   detector.Attach(actor2);
-  DALI_TEST_EQUALS(2u, detector.GetAttachedActorCount(), TEST_LOCATION);
+  DALI_TEST_EQUALS((size_t)2u, detector.GetAttachedActorCount(), TEST_LOCATION);
 
   // Attach another five actors
   std::vector<Actor> myActors;
@@ -405,35 +405,35 @@ int UtcDaliGestureDetectorGetAttachedActors(void)
     myActors.push_back(actor);
     detector.Attach(actor);
   }
-  DALI_TEST_EQUALS(7u, detector.GetAttachedActorCount(), TEST_LOCATION);
+  DALI_TEST_EQUALS((size_t)7u, detector.GetAttachedActorCount(), TEST_LOCATION);
 
   // Detach actor2
   detector.Detach(actor2);
-  DALI_TEST_EQUALS(6u, detector.GetAttachedActorCount(), TEST_LOCATION);
+  DALI_TEST_EQUALS((size_t)6u, detector.GetAttachedActorCount(), TEST_LOCATION);
 
   // Attach actor1 again, count should not increase.
   detector.Attach(actor1);
-  DALI_TEST_EQUALS(6u, detector.GetAttachedActorCount(), TEST_LOCATION);
+  DALI_TEST_EQUALS((size_t)6u, detector.GetAttachedActorCount(), TEST_LOCATION);
 
   // Detach actor2 again, count should not decrease.
   detector.Detach(actor2);
-  DALI_TEST_EQUALS(6u, detector.GetAttachedActorCount(), TEST_LOCATION);
+  DALI_TEST_EQUALS((size_t)6u, detector.GetAttachedActorCount(), TEST_LOCATION);
 
   // Detach actor1.
   detector.Detach(actor1);
-  DALI_TEST_EQUALS(5u, detector.GetAttachedActorCount(), TEST_LOCATION);
+  DALI_TEST_EQUALS((size_t)5u, detector.GetAttachedActorCount(), TEST_LOCATION);
 
   // Create scoped actor, actor should be automatically removed from the detector when it goes out
   // of scope.
   {
     Actor scopedActor = Actor::New();
     detector.Attach(scopedActor);
-    DALI_TEST_EQUALS(6u, detector.GetAttachedActorCount(), TEST_LOCATION);
+    DALI_TEST_EQUALS((size_t)6u, detector.GetAttachedActorCount(), TEST_LOCATION);
   }
-  DALI_TEST_EQUALS(5u, detector.GetAttachedActorCount(), TEST_LOCATION);
+  DALI_TEST_EQUALS((size_t)5u, detector.GetAttachedActorCount(), TEST_LOCATION);
 
   // Detach all so nothing remains.
   detector.DetachAll();
-  DALI_TEST_EQUALS(0u, detector.GetAttachedActorCount(), TEST_LOCATION);
+  DALI_TEST_EQUALS((size_t)0u, detector.GetAttachedActorCount(), TEST_LOCATION);
   END_TEST;
 }
